requirements.txt
django-restframework
pillow
django-crispy-forms
crispy-bootstrap5
python-dotenv
django-allauth
django-allauth[socialaccount]
stripe